pyproject.toml
[project]
name = "shippyapi"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi[standard]>=0.116.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pymongo>=4.14.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
]
